process/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
faiss-cpu>=1.7.4
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
joblib>=1.3.0

gradio>=4.42.0,<5
# audioop-lts not needed for Python 3.12 (audioop is built-in)
# audioop-lts>=0.2.1
Flask>=3.0.0
flask-cors>=4.0.0
